Fix intentional demo bug.

+++ src/main/java/com/acme/edu/primitives/Demo.java
@@ -61,7 +61,7 @@ public class Demo {
         byte incr = 0b111;
         System.out.println(incr--);
         byte bbbbb = incr++;
-        byte bbbbbb = incr + 1; //arith. promotion
+        byte bbbbbb = (byte)(incr + 1); //arith. promotion
 
         incr /= 2; //incr = incr + 2;
